requirements.txt
allure-pytest==2.13.5
pytest==8.3.3
pytest-playwright==0.6.2
playwright==1.48.0
